+++ Editor/UFolderUtility.cs
@@ -19,10 +19,10 @@ namespace NKStudio
         internal static bool IsExpanded(GameObject gameObject)
         {
             var expandedGameObjects = GetExpandedGameObjects();
-
+ 
             if (expandedGameObjects != null)
                 return expandedGameObjects.Contains(gameObject);
-
+ 
             return false;
         }
 
@@ -32,19 +32,104 @@ namespace NKStudio
         private static List<GameObject> GetExpandedGameObjects()
         {
             object sceneHierarchy = HierarchyWindowAdapter.GetFirstHierarchy();
-
+ 
             MethodInfo methodInfo = sceneHierarchy
                 .GetType()
                 .GetMethod("GetExpandedGameObjects");
-
+ 
             if (methodInfo != null)
             {
                 object result = methodInfo.Invoke(sceneHierarchy, Array.Empty<object>());
-
+ 
                 return (List<GameObject>)result;
             }
-
+ 
             return null;
+        }
+#endif
+
+#if UNITY_6000_5_OR_NEWER
+        /// <summary>
+        /// 게임 오브젝트 및 확장 상태에 맞춰 알맞은 폴더 아이콘을 반환합니다.
+        /// </summary>
+        public static Texture2D GetFolderIcon(GameObject obj, bool isExpanded)
+        {
+            if (obj.activeInHierarchy)
+            {
+                return GetFolderIconActive(obj, isExpanded);
+            }
+            else
+            {
+                return GetFolderIconDeActive(obj, isExpanded);
+            }
+        }
+
+        private static Texture2D GetFolderIconActive(GameObject obj, bool isExpanded)
+        {
+            int childCount = obj.transform.childCount;
+            string iconName;
+            bool hasChild = childCount > 0;
+            if (hasChild)
+            {
+                if (EditorGUIUtility.isProSkin)
+                {
+                    if (isExpanded)
+                        iconName = "FolderOpened On Icon";
+                    else
+                        iconName = "Folder On Icon";
+                }
+                else
+                {
+                    if (isExpanded)
+                        iconName = "FolderOpened Icon";
+                    else
+                        iconName = "Folder Icon";
+                }
+            }
+            else
+            {
+                if (EditorGUIUtility.isProSkin)
+                    iconName = "FolderEmpty On Icon";
+                else
+                    iconName = "FolderEmpty Icon";
+            }
+
+            GUIContent folderIconContent = EditorGUIUtility.IconContent(iconName);
+            return folderIconContent.image as Texture2D;
+        }
+
+        private static Texture2D GetFolderIconDeActive(GameObject obj, bool isExpanded)
+        {
+            int childCount = obj.transform.childCount;
+            string iconName;
+            bool hasChild = childCount > 0;
+            if (hasChild)
+            {
+                if (EditorGUIUtility.isProSkin)
+                {
+                    if (isExpanded)
+                        iconName = "FolderOpened Icon";
+                    else
+                        iconName = "Folder Icon";
+                }
+                else
+                {
+                    if (isExpanded)
+                        iconName = "FolderOpened On Icon";
+                    else
+                        iconName = "Folder On Icon";
+                }
+            }
+            else
+            {
+                if (EditorGUIUtility.isProSkin)
+                    iconName = "FolderEmpty Icon";
+                else
+                    iconName = "FolderEmpty On Icon";
+            }
+
+            GUIContent folderIconContent = EditorGUIUtility.IconContent(iconName);
+            return folderIconContent.image as Texture2D;
         }
 #endif
 
@@ -55,7 +140,7 @@ namespace NKStudio
             ;
             TextAsset packageJson = AssetDatabase.LoadAssetAtPath<TextAsset>(path);
             PackageInfo info = JsonUtility.FromJson<PackageInfo>(packageJson.text);
-
+ 
             Debug.Log($"UFolder v{info.version}");
         }